Move storybook component to dedicated stories directory; remove unused CSS class
