+++ frontend/src/app/testResults/submissions/Submissions.scss
@@ -1,11 +1,3 @@
-.sr-showing-results-on-page {
-  margin-left: 2rem;
-  color: #555;
-  font-size: 50%;
-  vertical-align: middle;
-  font-weight: normal;
-}
-
 .test-date-time-container {
   display: flex;
   justify-content: flex-start;
