+++ frontend/src/stories/testResults/Uploads.stories.tsx
@@ -2,10 +2,9 @@ import { Provider } from "react-redux";
 import { MemoryRouter } from "react-router-dom";
 import { Meta, Story } from "@storybook/react";
 
-import { store } from "../../store";
-import { StoryGraphQLProvider } from "../../../stories/storyMocks";
-
-import Uploads from "./Uploads";
+import { store } from "../../app/store";
+import { StoryGraphQLProvider } from "../../stories/storyMocks";
+import Uploads from "../../app/testResults/uploads/Uploads";
 
 type Props = {};
